산책하기랑 지도보기 액티비티 분리
