+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     package="com.example.petsi">
-
     <!-- 지도 및 날씨 관련 권한 설정 -->
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
@@ -18,25 +17,25 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Petsi"
         tools:targetApi="31">
-
+        <activity
+            android:name=".activity_watching_map"
+            android:exported="false" />
         <!-- ✅ MainActivity를 앱 시작 액티비티로 설정 -->
         <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <activity
             android:name=".activity_walking_with_map"
-            android:exported="true" />
-
-        <!-- ✅ 네이버 지도 API 키 -->
+            android:exported="true" /> <!-- ✅ 네이버 지도 API 키 -->
         <meta-data
             android:name="com.naver.maps.map.NCP_KEY_ID"
             android:value="k23b2ztztk" />
-
     </application>
+
 </manifest>
