+++ app/src/main/res/layout/activity_walking_with_map.xml
@@ -26,94 +26,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
 
-    <!-- ✅ 오버레이: 검색창 + 카테고리 버튼 묶음 -->
-    <FrameLayout
-        android:id="@+id/search_category_overlay"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        app:layout_constraintTop_toBottomOf="@id/customToolbar"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:elevation="15dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="8dp"
-            android:background="@android:color/transparent">
-
-            <!-- 🔍 검색 바 -->
-            <LinearLayout
-                android:id="@+id/layout_search"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:background="@android:color/white"
-                android:padding="8dp"
-                android:elevation="5dp">
-
-                <EditText
-                    android:id="@+id/et_search"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="1"
-                    android:hint="장소를 입력하세요"
-                    android:padding="8dp"
-                    android:background="@android:color/darker_gray"
-                    android:textColor="@android:color/black"
-                    android:textSize="14sp" />
-
-                <Button
-                    android:id="@+id/btn_search"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="검색"
-                    android:textColor="@android:color/white"
-                    android:background="#4CAF50"
-                    android:layout_marginStart="8dp" />
-            </LinearLayout>
-
-            <!-- 📍 카테고리 버튼 -->
-            <LinearLayout
-                android:id="@+id/layout_category_buttons"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:layout_marginTop="8dp"
-                android:gravity="center_horizontal"
-                android:layout_gravity="center_horizontal"
-                android:background="@android:color/transparent">
-
-                <Button
-                    android:id="@+id/btn_cafe"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="카페" />
-
-                <Button
-                    android:id="@+id/btn_food"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="음식점"
-                    android:layout_marginStart="8dp" />
-
-                <Button
-                    android:id="@+id/btnVet"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="동물병원"
-                    android:layout_marginStart="8dp" />
-                <Button
-                    android:id="@+id/btnPark"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="공원"
-                    android:layout_marginStart="8dp" />
-            </LinearLayout>
-        </LinearLayout>
-    </FrameLayout>
 
     <!-- ✅ 장소 정보 카드뷰 (초기엔 안보임) -->
     <androidx.cardview.widget.CardView
